Create `grading_system.dart` and move classes
